vultr: delete ipv4 if instance is gone
This commit fixes an issue where the provider errors out if the user
manually deletes an instance with attached IPv4 resources. In this case,
the provider would be unable to find the underlying instance, causing
the IPv4 state read to fail. Instead, since IPv4s must be associated
with an instance and are deleted when the instance is destroyed, assume
that the resources are gone if the instance cannot be found.