vultr: fix OS ID for snapshots
The Vultr API models the the OS of an instance as being
dependent on _two_ things: the OS selected at creation AND the snapshot
selected at creation. This turns out to be the case as well for
applications, i.e. if a chosen snapshot is of an application server,
then once the instance boots, its application ID will be that of the
original snapshotted app.

To solve this in this provider:
the `os_id` field of the `vultr_instance` resource becomes optional
and is computed by the provider. When creating an instance based on a
snapshot, the user must not supply an `os_id`. Once the instance is up,
the `os_id` field is computed and set to the OS ID of the machine that
the snapshot was based on. This is a slightly larger public API change
but requires the user to configure fewer options.

This option simplifies the interface of the provider
and allows the user to do less work. From now on, when booting a
snapshot, only the `snapshot_id` must be configured. The `os_id` will be
computed.