Implements 'vultr_reverse_name' resource
It works for IPv4 and IPv6 likewise.
An extra resource has been chosen, because in order to set the
reverse name, the IP(s) of the instance have to be known. And
that's only after they have been created.

When an IPv4 reverse name is deleted, nothing really happens.
This is because there is not API of Vultr to 'reset' the reverse
name.